requirements.txt
fastapi
uvicorn[standard]
openai
edge-tts
aiofiles
python-dotenv
